Replace graph with a simple table lookup
Instead of using a graph, we can use a table lookup to determine a list
of possible reactions where the two compounds are present. We can apply
the same "post-processing" beforehand and sort the lists, so that we can
simply return the top `max_rank` reactions. Doing so returns a list of
the same size, with probabilities in the same location, thus for a
fixed seed the result is the same before and after this commit.

We still need to change a snapshot test, but only for ease of
comprehension. The "weighted reaction rankings" test expects the data to
be a dictionary, where:
1. There is a list of compounds for each reaction. We remove this
   because this list is not used and can be easily obtained by just
   looking at the reaction. `components = set(e.reac) | set(e.prod)`
2. While the old could would do a multiplicative inverse after the
   function returns, we do so in the pre-generated data files instead.
   Additionally, in the new version, we return a normalised list of
   probabilities (ie, they sum to 1).

I could fake the old functionality, but that seems to be a lot of work
and technicaly debt for no good reason.